intial commit, made migration struct configurable + public
